requirements.txt
chromadb
sentence-transformers
torch
torchvision
ftfy
regex
tqdm
Pillow
shutil
argparse
pypdf
streamlit